backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart
pydantic>=2.0
pydantic-settings

# Redis
redis>=5.0.0

# STT
faster-whisper>=1.0.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.24.0

# LLM — Groq
groq>=0.5.0
httpx>=0.27.0
tenacity>=8.2.0

# LLM — Gemini Vision
google-generativeai>=0.7.0

# TTS
kokoro>=0.9.0

# Image processing
pillow>=10.0.0

# Utils
python-dotenv
aiofiles
